gpg hook: Make signature verification optional
Make signature verification optional.  When disabled, the hook only
enforces that a signature is present but does not necessarily verify its
validity.

The used configuration key uses a default value of 'gentoo-devs'.  This
is meant to match the gkeys seed group, to make the configuration
future-proof for gkeys-based verification with multiple possible groups.